requirements.txt
streamlit
pandas
numpy
plotly
yfinance
python-bcb
fpdf2
kaleido
